eatgo-api/src/main/java/com/codingcat/eatgo/application/RestaurantService.java
package com.codingcat.eatgo.application;

import com.codingcat.eatgo.domain.MenuItem;
import com.codingcat.eatgo.domain.MenuItemRepository;
import com.codingcat.eatgo.domain.Restaurant;
import com.codingcat.eatgo.domain.RestaurantRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class RestaurantService {

    private final RestaurantRepository restaurantRepository;
    private final MenuItemRepository menuItemRepository;

    public RestaurantService(RestaurantRepository restaurantRepository, MenuItemRepository menuItemRepository) {
        this.restaurantRepository = restaurantRepository;
        this.menuItemRepository = menuItemRepository;
    }

    public Restaurant getRestaurantById(Long id){
        Restaurant restaurant = restaurantRepository.findById(id);
        List<MenuItem> menuItems = menuItemRepository.findAllByRestaurantId(id);
        restaurant.setMenuItems(menuItems);

        return restaurant;
    }

    public List<Restaurant> getRestaurants() {
        return  restaurantRepository.findAll();
    }
}
